fix(orchestration): move ToolCallSummary to always-compiled scheduler module
crate::scheduler's TaskOutcome/SchedulerAction variants carry
Option<Vec<ToolCallSummary>> unconditionally, but ToolCallSummary was
defined inside verifier.rs, itself gated behind #[cfg(feature =
"llm-planning")]. zeph-orchestration failed to compile with
llm-planning off (caught by cargo-hack's --each-feature isolation
check, #6744).

Moves the struct definition into scheduler/mod.rs (unconditional,
and its primary consumer already); verifier.rs re-exports it via
`pub use super::scheduler::ToolCallSummary` so every existing
`crate::verifier::ToolCallSummary` import keeps resolving unchanged.
lib.rs's public re-export moves from the llm-planning-gated line to
the unconditional scheduler:: one.
